extract abstract class Login
