+++ lib/features/login/data/login.dart
@@ -1,0 +1,5 @@
+import 'login_result.dart';
+
+abstract class Login {
+  Future<LoginResult> login(String username, String password);
+}
