--- lib/features/login/data/login_api/login_use_case_errors.dart
@@ -1,1 +1,0 @@
-enum LoginUseCaseErrors { connectivity, invalidData, invalidCredential }
