+++ lib/features/login/data/remote_login/login_api_errors.dart
@@ -1,0 +1,1 @@
+enum LoginApiErrors { connectivity, invalidData, invalidCredential }
