+++ lib/features/login/data/remote_login/remote_login_api.dart
@@ -2,15 +2,17 @@ import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 
-import 'login_result.dart';
-import 'login_use_case_errors.dart';
+import '../login.dart';
+import '../login_result.dart';
+import 'login_api_errors.dart';
 
-class RemoteLoginApi {
+class RemoteLoginApi extends Login {
   RemoteLoginApi(this._client, this._url);
 
   final http.Client _client;
   final Uri _url;
 
+  @override
   Future<LoginResult> login(String username, String password) async {
     try {
       final http.Response response = await _client.post(
@@ -23,12 +25,12 @@ class RemoteLoginApi {
       if (response.statusCode == 200) {
         return _tryParse(response.body);
       } else if (response.statusCode == 401) {
-        return Future<LoginResult>.error(LoginUseCaseErrors.invalidCredential);
+        return Future<LoginResult>.error(LoginApiErrors.invalidCredential);
       } else {
-        return Future<LoginResult>.error(LoginUseCaseErrors.invalidData);
+        return Future<LoginResult>.error(LoginApiErrors.invalidData);
       }
     } catch (error) {
-      return Future<LoginResult>.error(LoginUseCaseErrors.connectivity);
+      return Future<LoginResult>.error(LoginApiErrors.connectivity);
     }
   }
 
@@ -36,7 +38,7 @@ class RemoteLoginApi {
     try {
       return Future<LoginResult>.value(LoginResult.fromJson(jsonDecode(body)));
     } catch (error) {
-      return Future<LoginResult>.error(LoginUseCaseErrors.invalidData);
+      return Future<LoginResult>.error(LoginApiErrors.invalidData);
     }
   }
 }
